requirements.txt
fastapi>=0.110
uvicorn>=0.27
tree-sitter>=0.21
tree-sitter-typescript>=0.21
tree-sitter-javascript>=0.21
jinja2>=3.1
pyyaml>=6.0
